+++ Magic/CHANGELOG.md
@@ -11,6 +11,7 @@
  - Add `clear_containers` parameter to engineering spells, to clear (the copied versions of) containers when copying them
  - Add `min_radius` parameter to AreaOfEffect action, to select entities within a ring
  - Add `/mage discover` command for learning crafting recipes
+ - Add `/mitem amount` command to change the amount of an item stack
 
 ## Changes